Update teku_bn.sh
This is for disabling rate limit to let evil teku serve RPC without rate limiting peers.
I think This option should be better integrate with the hive test suite, so it can be enabled or disabled based on the specific scenario. I don't feel it is correct to have disabled peer rate limit by default for teku.